# Conflicts:
#	XenAdmin/Wizards/ImportWizard/ImportWizard.cs
#	XenModel/Actions/VM/VMCopyAction.cs

+++ XenAdmin/Controls/SectionHeaderLabel.cs
@@ -300,7 +300,7 @@ namespace XenAdmin.Controls
                     }
                 case VerticalAlignment.Middle:
                     {
-                        Y = (this.Height - this.Padding.Vertical) / 2;
+                        Y = ((float) this.Height - this.Padding.Vertical) / 2;
                         break;
                     }
                 case VerticalAlignment.Bottom:
